samples/python/src/microsoft_agent_framework/hybrid_tool_calling/requirements.txt
# Microsoft Agent Framework (Preview)
agent-framework>=0.1.0a1

# Azure Identity for authentication
azure-identity>=1.15.0

# Environment variable loading
python-dotenv>=1.0.0
